Added basic (.ply) file maker from 2d images
I added the data for the small dataset. Worked out the most basic appraoch possible to see if it works. It does. The current output is in (.bin) and then converted to (.ply). The model is sparse and doen't look the best when visualized, but there is enough information to extract heights and lenghts.

visualization was run from terminal